+++ pkgs/ffigen/lib/src/code_generator/func.dart
@@ -145,13 +145,15 @@ class Func extends LookUpBinding {
     }
 
     if (ffiNativeConfig.enabled) {
-      final assetString = ffiNativeConfig.assetId != null
-          ? ", assetId: '${ffiNativeConfig.assetId}'"
-          : '';
-      final isLeafString = isLeaf ? ', isLeaf:true' : '';
       final nativeFuncName = needsWrapper ? funcVarName : enclosingFuncName;
       s.write('''
-@${w.ffiLibraryPrefix}.Native<$cType>(symbol: '$originalName'$assetString$isLeafString)
+${makeNativeAnnotation(
+        w,
+        nativeType: cType,
+        dartName: nativeFuncName,
+        nativeSymbolName: originalName,
+        isLeaf: isLeaf,
+      )}
 external $ffiReturnType $nativeFuncName($ffiArgDeclString);
 
 ''');
@@ -163,6 +165,15 @@ external $ffiReturnType $nativeFuncName($ffiArgDeclString);
 $dartReturnType $enclosingFuncName($libArg$dartArgDeclString) => $funcImplCall;
 
 ''');
+      }
+
+      if (exposeSymbolAddress) {
+        // Add to SymbolAddress in writer.
+        w.symbolAddressWriter.addNativeSymbol(
+          type:
+              '${w.ffiLibraryPrefix}.Pointer<${w.ffiLibraryPrefix}.NativeFunction<$cType>>',
+          name: name,
+        );
       }
     } else {
       funcPointerName = w.wrapperLevelUniqueNamer.makeUnique('_${name}Ptr');