+++ pkgs/ffigen/lib/src/code_generator/global.dart
@@ -2,9 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import '../config_provider/config_types.dart';
 import 'binding.dart';
 import 'binding_string.dart';
 import 'compound.dart';
+import 'pointer.dart';
 import 'type.dart';
 import 'utils.dart';
 import 'writer.dart';
@@ -22,6 +24,7 @@ import 'writer.dart';
 class Global extends LookUpBinding {
   final Type type;
   final bool exposeSymbolAddress;
+  final FfiNativeConfig nativeConfig;
   final bool constant;
 
   Global({
@@ -32,6 +35,7 @@ class Global extends LookUpBinding {
     super.dartDoc,
     this.exposeSymbolAddress = false,
     this.constant = false,
+    this.nativeConfig = const FfiNativeConfig(enabled: false),
   });
 
   @override
@@ -41,35 +45,63 @@ class Global extends LookUpBinding {
     if (dartDoc != null) {
       s.write(makeDartDoc(dartDoc!));
     }
-    final pointerName = w.wrapperLevelUniqueNamer.makeUnique('_$globalVarName');
     final dartType = type.getFfiDartType(w);
     final cType = type.getCType(w);
 
-    s.write(
-        "late final ${w.ffiLibraryPrefix}.Pointer<$cType> $pointerName = ${w.lookupFuncIdentifier}<$cType>('$originalName');\n\n");
-    final baseTypealiasType = type.typealiasType;
-    if (baseTypealiasType is Compound) {
-      if (baseTypealiasType.isOpaque) {
-        s.write(
-            '${w.ffiLibraryPrefix}.Pointer<$cType> get $globalVarName => $pointerName;\n\n');
-      } else {
-        s.write('$dartType get $globalVarName => $pointerName.ref;\n\n');
+    if (nativeConfig.enabled) {
+      if (type case final ConstantArray arr) {
+        s.writeln(makeArrayAnnotation(w, arr));
+      }
+
+      s
+        ..writeln(makeNativeAnnotation(
+          w,
+          nativeType: cType,
+          dartName: globalVarName,
+          nativeSymbolName: originalName,
+          isLeaf: false,
+        ))
+        ..write('external ');
+      if (constant) {
+        s.write('final ');
+      }
+
+      s.writeln('$dartType $globalVarName;\n');
+
+      if (exposeSymbolAddress) {
+        w.symbolAddressWriter.addNativeSymbol(
+            type: '${w.ffiLibraryPrefix}.Pointer<$cType>', name: name);
       }
     } else {
-      s.write('$dartType get $globalVarName => $pointerName.value;\n\n');
-      if (!constant) {
-        s.write(
-            'set $globalVarName($dartType value) => $pointerName.value = value;\n\n');
-      }
-    }
+      final pointerName =
+          w.wrapperLevelUniqueNamer.makeUnique('_$globalVarName');
 
-    if (exposeSymbolAddress) {
-      // Add to SymbolAddress in writer.
-      w.symbolAddressWriter.addSymbol(
-        type: '${w.ffiLibraryPrefix}.Pointer<$cType>',
-        name: name,
-        ptrName: pointerName,
-      );
+      s.write(
+          "late final ${w.ffiLibraryPrefix}.Pointer<$cType> $pointerName = ${w.lookupFuncIdentifier}<$cType>('$originalName');\n\n");
+      final baseTypealiasType = type.typealiasType;
+      if (baseTypealiasType is Compound) {
+        if (baseTypealiasType.isOpaque) {
+          s.write(
+              '${w.ffiLibraryPrefix}.Pointer<$cType> get $globalVarName => $pointerName;\n\n');
+        } else {
+          s.write('$dartType get $globalVarName => $pointerName.ref;\n\n');
+        }
+      } else {
+        s.write('$dartType get $globalVarName => $pointerName.value;\n\n');
+        if (!constant) {
+          s.write(
+              'set $globalVarName($dartType value) => $pointerName.value = value;\n\n');
+        }
+      }
+
+      if (exposeSymbolAddress) {
+        // Add to SymbolAddress in writer.
+        w.symbolAddressWriter.addSymbol(
+          type: '${w.ffiLibraryPrefix}.Pointer<$cType>',
+          name: name,
+          ptrName: pointerName,
+        );
+      }
     }
 
     return BindingString(type: BindingStringType.global, string: s.toString());