+++ pkgs/ffigen/lib/src/header_parser/sub_parsers/var_parser.dart
@@ -26,17 +26,15 @@ Global? parseVarDeclaration(clang_types.CXCursor cursor) {
   _logger.fine('++++ Adding Global: ${cursor.completeStringRepr()}');
 
   final cType = cursor.type();
-  final type = cType.toCodeGenType();
+
+  final type = cType.toCodeGenType(
+      // Native fields can be arrays, but if we use the lookup based method of
+      // reading fields there's no way to turn a Pointer into an array.
+      supportNonInlineArray: config.ffiNativeConfig.enabled);
   if (type.baseType is UnimplementedType) {
     _logger.fine('---- Removed Global, reason: unsupported type: '
         '${cursor.completeStringRepr()}');
     _logger.warning("Skipped global variable '$name', type not supported.");
-    return null;
-  }
-
-  if (config.ffiNativeConfig.enabled) {
-    _logger
-        .warning("Skipped global variable '$name', not supported in Natives.");
     return null;
   }
 
@@ -46,9 +44,11 @@ Global? parseVarDeclaration(clang_types.CXCursor cursor) {
     usr: usr,
     type: type,
     dartDoc: getCursorDocComment(cursor),
-    exposeSymbolAddress: config.functionDecl.shouldIncludeSymbolAddress(name),
+    exposeSymbolAddress: config.globals.shouldIncludeSymbolAddress(name),
     constant: cType.isConstQualified,
+    nativeConfig: config.ffiNativeConfig,
   );
   bindingsIndex.addGlobalVarToSeen(usr, global);
+
   return global;
 }